data-privacy-vault/data_privacy_vault/auth.py
from data_privacy_vault import config
from flask import request, Response
from functools import wraps

USER = config.auth["username"]
PASSWORD = config.auth["password"]

def check_auth(username, password):
    return username == USER and password == PASSWORD

def authenticate():
    """Sends a 401 response that enables basic auth"""
    return Response(
    'Could not verify your access level for this URL.\n'
    'You have to login with proper credentials', 401,
    {'WWW-Authenticate': 'Basic realm="Login Required"'})

def requires_auth(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        auth = request.authorization
        if not auth or not check_auth(auth.username, auth.password):
            return authenticate()
        return f(*args, **kwargs)
    return decorated
